+++ qoqo_qiskit_devices/src/devices/ibm_belem.rs
@@ -35,6 +35,9 @@ impl IBMBelemDeviceWrapper {
     /// Create a new IBMBelemDevice instance.
     #[new]
     pub fn new() -> Self {
+        Python::with_gil(|py| {
+            py.run("import warnings; warnings.warn(\"Device ibmq_belem has been retired. Setting noise information is not possible.\", stacklevel=2)", None, None).unwrap();
+        });
         Self {
             internal: IBMBelemDevice::new(),
         }