+++ qoqo_qiskit_devices/python_tests/test_initialization.py
@@ -21,50 +21,61 @@ from qoqo_qiskit_devices import ibm_devices
 def test_belem():
     """Test IBMBelemDevice initialization."""
     with warnings.catch_warnings(record=True) as w:
-        belem = ibm_devices.IBMBelemDevice()
+        ibm_devices.IBMBelemDevice()
         assert len(w) == 1
+        assert issubclass(w[-1].category, DeprecationWarning)
+        assert "retired" in str(w[-1].message)
 
 
 def test_jakarta():
     """Test IBMJakartaDevice initialization."""
     with warnings.catch_warnings(record=True) as w:
-        jakarta = ibm_devices.IBMJakartaDevice()
+        ibm_devices.IBMJakartaDevice()
         assert len(w) == 1
+        assert issubclass(w[-1].category, DeprecationWarning)
+        assert "retired" in str(w[-1].message)
+
 
 def test_lagos():
     """Test IBMLagosDevice initialization."""
-    lagos = ibm_devices.IBMLagosDevice()
+    ibm_devices.IBMLagosDevice()
 
 
 def test_lima():
     """Test IBMLimaDevice initialization."""
     with warnings.catch_warnings(record=True) as w:
-        lima = ibm_devices.IBMLimaDevice()
+        ibm_devices.IBMLimaDevice()
         assert len(w) == 1
+        assert issubclass(w[-1].category, DeprecationWarning)
+        assert "retired" in str(w[-1].message)
 
 
 def test_manila():
     """Test IBMManilaDevice initialization."""
     with warnings.catch_warnings(record=True) as w:
-        manila = ibm_devices.IBMManilaDevice()
+        ibm_devices.IBMManilaDevice()
         assert len(w) == 1
+        assert issubclass(w[-1].category, DeprecationWarning)
+        assert "retired" in str(w[-1].message)
 
 
 def test_nairobi():
     """Test IBMNairobiDevice initialization."""
-    nairobi = ibm_devices.IBMNairobiDevice()
+    ibm_devices.IBMNairobiDevice()
 
 
 def test_perth():
     """Test IBMPerthDevice initialization."""
-    perth = ibm_devices.IBMPerthDevice()
+    ibm_devices.IBMPerthDevice()
 
 
 def test_quito():
     """Test IBMQuitoDevice initialization."""
     with warnings.catch_warnings(record=True) as w:
-        quito = ibm_devices.IBMQuitoDevice()
+        ibm_devices.IBMQuitoDevice()
         assert len(w) == 1
+        assert issubclass(w[-1].category, DeprecationWarning)
+        assert "retired" in str(w[-1].message)
 
 
 if __name__ == "__main__":