+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This changelog tracks changes of the qoqo_qiskit project starting at version 0.1
 
 * Updated to qoqo 1.7
 * Added deprecation warnings for retired devices (Belem, Jakarta, Lima, Manila, Quito)
+* Added `Identity` gate error info retrieval
 
 ### 0.1.10